+++ src/views/travisStatusBarItem.ts
@@ -1,18 +1,17 @@
 import * as _ from 'lodash';
 import { ExtensionContext, StatusBarAlignment, StatusBarItem, window, workspace } from 'vscode';
-import { ProjectDetails } from '../common/ProjectDetails';
-import { ActiveRepository } from '../common/ActiveRepository';
+import { ActiveRepositorySingleton } from '../common/ActiveRepositorySingleton';
 import { buildsURLTemplate } from '../common/apiTemplates';
 
 const axios = require('axios');
 
 export class TravisStatusBar {
   private _statusBarItem: StatusBarItem;
-  private activeRepositoryInstance: any;
+  private singleton: any;
 
   constructor(private context: ExtensionContext) {
     if (workspace && workspace.rootPath) {
-      this.activeRepositoryInstance = new ActiveRepository(this.context, workspace.rootPath);
+      this.singleton = ActiveRepositorySingleton.getInstance();
     }
     this._statusBarItem = window.createStatusBarItem(StatusBarAlignment.Left, 100);
     this.subscribe();
@@ -20,25 +19,23 @@ export class TravisStatusBar {
   }
 
   async loadStatusBarData() {
-    const headers = ActiveRepository.headers(this.context);
+    const headers = this.singleton.headers();
 
-    let activeRepo = {id: await this.activeRepositoryInstance.repositoryId};
-    if (!activeRepo || !activeRepo.id) {
-      activeRepo = await this.activeRepositoryInstance.getActiveRepositoryDetails();
-    }
+    const activeRepo = {id: await this.singleton.repositoryId()};
 
     const response = await axios.get(
       buildsURLTemplate({
-        base: ProjectDetails.getProjectDetails(this.context).base,
+        base: this.singleton.base(),
         repoId: _.get(activeRepo, 'id'),
-        branch: this.activeRepositoryInstance.branch
+        branch: this.singleton.branch()
       }),
     {headers: headers}
     );
 
     const build: any = _.head(_.get(response, 'data.builds'));
-    const statusData = await {id: build.id, name: this.activeRepositoryInstance.branch, state: build.state};
+    const statusData = await {id: build.id, repoName: this.singleton.repository(), branchName: this.singleton.branch(), state: build.state};
     return statusData;
+
   }
 
   public getStatusIcon(state: string) {
@@ -52,23 +49,30 @@ export class TravisStatusBar {
       case 'failed':
         return 'x';
       case 'errored':
-        return 'circle-slash';
+        return 'x';
       default:
         return '';
     }
   }
 
   async updateStatusBar(refresh: boolean = false) {
-    if (ProjectDetails.isTravisProject()) {
+    if (this.singleton && this.singleton.isTravisProject()) {
       if (refresh) {
-        this._statusBarItem.text = `Travis: $(sync~spin)`;
-        this._statusBarItem.tooltip = `refreshing`;
+        this._statusBarItem.text = 'Travis: $(sync~spin)';
+        this._statusBarItem.tooltip = 'refreshing..';
         this.show();
       }
-      const data: any = await this.loadStatusBarData();
-      this._statusBarItem.text = `Travis: ${data.name}: $(${this.getStatusIcon(data.state)})`;
-      this._statusBarItem.tooltip = `${data.name}: ${data.id} - ${data.state}`;
-      this.show();
+
+      try {
+        const data: any = await this.loadStatusBarData();
+        this._statusBarItem.text = `Travis: ${data.repoName}: $(${this.getStatusIcon(data.state)})`;
+        this._statusBarItem.tooltip = `${data.branchName}: ${data.id} - ${data.state}`;
+        this.show();
+      }
+      catch (e) {
+        console.error(e);
+        this.hide();
+      }
     }
   }